Work Done: --------- - Remove vendor folder from .gitignore. - Integrate with laravel-jsend for unifying the sent response. - Update all the already implemented response to use jsend. - Create new migration file for followers table. - Add the needed relation for followers and followings into User model. - Create api for follow and unfollow user. - Add validations to prevent follow not existing user, follow the logged in user and allow only authenticated users. - Add tests for testing follow/unfollow feature. - Return error if user try to delete non-exist tweet. - Integrate with phpunit-arraysubset-asserts to check if the response contains array.

+++ app/Http/Controllers/TweetController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Http\Requests\TweetStoreRequest;
 use App\Models\tweet;
 use App\Repositories\Repository;
-use Illuminate\Http\Request;
 
 class TweetController extends Controller
 {
@@ -18,26 +17,28 @@ class TweetController extends Controller
         $this->model = new Repository($tweet);
     }
 
-    public function store(TweetStoreRequest $request){
+    public function store(TweetStoreRequest $request)
+    {
         $data = $request->all();
         $validated = $request->validated();
-        if($validated) {
+        if ($validated) {
             $data['user_id'] = auth()->user()->id;
             $tweet = $this->model->create($data);
-            return response()->json(['success' =>  'Tweet successfully created!', 'tweet' => $tweet], 200);
-
+            return jsend_success($tweet);
         }
-        return $validated;
     }
 
-    public function delete($id){
+    public function delete($id)
+    {
         $tweet = $this->model->find($id);
-        if(auth()->user()->id == $tweet->user_id){
-            $this->model->delete($id);
-            return response()->json(['success' =>  'Tweet successfully deleted!', 'code' => 200], 200);
-        }else {
-            return response()->json(['errors' =>  'Permission Denied! You are not authorized to preform this action.',
-                'code' => 402], 402);
+        if ($tweet) {
+            if (auth()->user()->id == $tweet->user_id) {
+                $this->model->delete($id);
+                return jsend_success(['message' => 'Tweet successfully deleted!']);
+            } else {
+                return jsend_fail(['errors' => 'Permission Denied! You are not authorized to preform this action.']);
+            }
         }
+        return jsend_fail(['errors' => 'The tweet your are trying to delete not exist!']);
     }
 }
